requirements.txt
# Web deployment requirements pro Streamlit Cloud / Railway
# Aktualizované pro Python 3.13 kompatibilitu
streamlit>=1.29.0
pandas>=2.2.0
plotly>=5.18.0
requests>=2.31.0
python-dotenv>=1.0.0
openpyxl>=3.1.2
xlsxwriter>=3.1.9
python-dateutil>=2.8.2
pytz>=2023.3
psycopg2-binary>=2.9.9
flask>=3.0.0
gspread>=6.0.0
google-auth>=2.27.0
